Fix issue #23

+++ src/framework/core/logger.h
@@ -39,7 +39,7 @@ class Logger
         MAX_LOG_HISTORY = 1000
     };
 
-    typedef std::function<void(Fw::LogLevel, const std::string&, int)> OnLogCallback;
+    typedef std::function<void(Fw::LogLevel, const std::string&, int64)> OnLogCallback;
 
 public:
     void log(Fw::LogLevel level, const std::string& message);
